Move `cython` requirement to Python package
This was previously in the C++ package where it was not needed. Also the
Python package was relying on the C++ package to supply `cython`
implicitly. This corrects that issue by making `cython` an explicit
dependency of the Python package where it is needed. Also drops it from
the C++ package where it is not needed.