Cascaded dispatch for type-erased API (#1711)
This enhancement PR provides the C++ infrastructure for a type-erased API to be exposed to C-only, Cython, and potentially other glue languages. This will be the first in a series of such API PRs to expose cugraph functionality to other languages.

Authors:
  - Andrei Schaffer (https://github.com/aschaffer)
  - Mark Harris (https://github.com/harrism)
  - AJ Schmidt (https://github.com/ajschmidt8)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Kumar Aatish (https://github.com/kaatish)

URL: https://github.com/rapidsai/cugraph/pull/1711