Fix Louvain hang in multi-GPU testing (#2028)
If the CSR + DCSR hybrid format is used to store a graph, the size of the offset array can be smaller than the number of vertices + 1. The previous code didn't consider this in calling cub's segmented sort. This PR fixes this mistake.

Authors:
  - Seunghwa Kang (https://github.com/seunghwak)

Approvers:
  - Chuck Hastings (https://github.com/ChuckHastings)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/2028