Allow class types to be properly represented in the `experimental_warning_wrapper()` return value (#2048)
`experimental_warning_wrapper()` returned a wrapper function that prints a `PendingDeprecationWarning` when an `experimental` API is used. It can be used for experimental functions and classes, however for classes, the wrapper it returns was a `function` type instead of a `type` type.  This caused the following problem:
```
>>> from cugraph.experimental import PropertyGraph
>>> pG = PropertyGraph()
>>> isinstance(pG, PropertyGraph)
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
TypeError: isinstance() arg 2 must be a type or tuple of types
```

This fix identifies the object to be wrapped and either creates a standard function wrapper for `function` types, or a class wrapper for `type` types:
```
>>> from cugraph.experimental import PropertyGraph
>>> pG = PropertyGraph()
>>> isinstance(pG, PropertyGraph)
True
```

NOTE: the `api_tools.py` module is copied in both cugraph and pylibcugraph. When cugraph starts using pylibcugraph, the copy in cugraph can be removed (see FIXME).

Authors:
  - Rick Ratzel (https://github.com/rlratzel)

Approvers:
  - Joseph Nke (https://github.com/jnke2016)
  - Don Acosta (https://github.com/acostadon)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/2048