Fix uniform neighborhood sampling remove duplicates (#2301)
@jnke2016 observed that the recent sampling code can return duplicates.

After some discussion it was decided to remove the duplicates and return a count of how many duplicates were removed.  The C++ implementation is updated in this PR to do that.  The C API will remain the same but is ready to support it once we have stopped using the old implementation.

Closes #2226

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/2301