Fix pandas SettingWithCopyWarning, which really shouldn't be ignored. (#2447)
See: https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy

This implementation requires column names to be strings, which is hopefully fine (but could be changed).

The downside is that this sometimes copies data when unnecessary.  It's pretty hard to know when objects are views or not when using public APIS.  Pandas objects have a `._is_view` attribute that tell us this.  I don't know about cugraph, or if this is even an issue in cugraph.

CC @rlratzel

Authors:
  - Erik Welch (https://github.com/eriknw)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/2447