Small fix to the MG PyG Test to Account for Current Sampling Behavior (#2666)
This PR will correct an error that shows up occasionally in the MG tests, especially when running on an odd number of GPUs.  This error is related to issue #2665 and resolves it temporarily while the C++ team works on a permanent fix.

Authors:
  - Alex Barghi (https://github.com/alexbarghi-nv)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/2666