Define a C API for data masking (#2630)
Add a C API for data masking.

Closes #2571 

The objective here is to add a C API that will allow us to associated a data mask with a graph.  Basic functions:
* Create a data mask
* Destroy a data mask
* Associated a data mask with a graph
* Disassociated a data mask from a graph
* Discover what data mask is associated with a graph

Current plan is to allow for the creation of a mask by the caller (the python layer for our current use case) and have it passed in.  We may later add support for other techniques.

This PR can be approved independently, but implementation of the API is dependent on #2679

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Naim (https://github.com/naimnv)

URL: https://github.com/rapidsai/cugraph/pull/2630