[HOTFIX] Removed unused testing code that caused a cudf dependency (#3065)
closes #3064 
closes #3057 

* Removed unused testing code that caused a `cudf` dependency
* Updated setup.py to remove cudf dependency
* Minor update to docstring to not specify cudf type
* Changed `cupy` dependency in `cugraph-service-server` to match the `cupy-cuda11x` dependency in `cugraph`
* Updated conda environment .yml files to add cupy

Note that the `pylibcugraph.testing` package is part of every install, and therefore any dependencies added to `testing` will cause all users to require them. The separate `extras_requires` are optional dependencies, in this case for running the unit tests. Some unit tests still use `cudf`, so cudf should be in `extras_require` and not `install_requires` since the ability to run the separate unit tests is optional.

Authors:
   - Rick Ratzel (https://github.com/rlratzel)

Approvers:
   - Brad Rees (https://github.com/BradReesWork)
   - Ray Douglass (https://github.com/raydouglass)