Adds better reporting of server subprocess errors during testing (#3012)
closes #3011 

This PR adds a check for the server subprocess exitcode while waiting for it to start so it fails faster. Since it can fail faster, it also increases the wait to a full 60 seconds, which has shown to be helpful when starting a MG server that connects to a dask cluster.

It also includes the subprocess stdout/stderr output in the exception raised or to the console depending on how it exited to help debug why the server didn't start or crashed.

Authors:
  - Rick Ratzel (https://github.com/rlratzel)

Approvers:
  - Brad Rees (https://github.com/BradReesWork)
  - Vibhu Jawa (https://github.com/VibhuJawa)
  - Alex Barghi (https://github.com/alexbarghi-nv)

URL: https://github.com/rapidsai/cugraph/pull/3012