Bump pinned pip wheel deps to 23.4 (#3253)
We introduced a change to pin RAPIDS wheel dependencies to the same release version. However, branch 23.04 was created before that last PR was merged, so as of now cugraph's 23.4 wheels are installing 23.2 RAPIDS dependencies. This PR updates those pins to the current release.

This should solve the error in #3160 (where the old pylibcugraph-cu11==23.2 is being pulled in the test step as opposed to the desired 23.4 version).

Authors:
  - Sevag H (https://github.com/sevagh)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)

URL: https://github.com/rapidsai/cugraph/pull/3253