Properly retrieve the dask worker from client calls (#3379)
A Dask PR which was merged last week changed the way workers are retrieved from client calls. This [PR](https://github.com/rapidsai/cugraph/pull/3361) assumed that the worker could be retrieve through `get_client()` and since the Dask version was pinned in RAFT to no break CI, it temporarily worked. But after the `dask `version was unpinned, this resulted in the issue this PR is closing. 

This PR leverages the `dask_worker` input argument which is populated with the worker when calling `client.run`.

closes #3378

Authors:
  - Joseph Nke (https://github.com/jnke2016)

Approvers:
  - Rick Ratzel (https://github.com/rlratzel)
  - Brad Rees (https://github.com/BradReesWork)

URL: https://github.com/rapidsai/cugraph/pull/3379