Changes to support gtest version 1.11 (#3511)
gtest version 1.11 adds a feature that complains if we define a test but do not instantiate it.  We have a few cases where we define a case, but we don't instantiate it at the moment.  This PR will ifdef these out so that they will run properly.

Also shrunk the size of an MG rmat test so that it would fit in a 1 GPU test.

Authors:
  - Chuck Hastings (https://github.com/ChuckHastings)

Approvers:
  - Naim (https://github.com/naimnv)
  - Seunghwa Kang (https://github.com/seunghwak)

URL: https://github.com/rapidsai/cugraph/pull/3511